Admin login page fix

+++ php/delete.php
@@ -33,12 +33,13 @@
 
 <?php
 include 'dblogin.php';
-
+if (($_SESSION['admin']) != "" && ($_SESSION['admin']) == "1" )
+{
 $delete_topic = strip_tags($_GET['id']);
 mysql_query("DELETE FROM topics WHERE id ='$delete_topic'") or die (mysql_error());
 
 echo "Topic Deleted";
-
+}
 ?>
